+++ swift/extractor/SwiftExtractor.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "SwiftExtractorConfiguration.h"
+#include "swift/extractor/SwiftExtractorConfiguration.h"
 #include <swift/AST/SourceFile.h>
 #include <swift/Frontend/Frontend.h>
 #include <memory>
