+++ swift/extractor/infra/BUILD.bazel
@@ -1,0 +1,10 @@
+load("//swift:rules.bzl", "swift_cc_library")
+
+swift_cc_library(
+    name = "infra",
+    hdrs = glob(["*.h"]),
+    visibility = ["//swift:__subpackages__"],
+    deps = [
+        "//swift/extractor/trap",
+    ],
+)
