+++ swift/extractor/infra/SwiftDispatcher.h
@@ -2,11 +2,12 @@
 
 #include "swift/extractor/trap/TrapArena.h"
 #include "swift/extractor/trap/TrapLabelStore.h"
+#include "swift/extractor/trap/TrapOutput.h"
 #include "swift/extractor/trap/generated/TrapClasses.h"
-#include "swift/extractor/SwiftTagTraits.h"
-#include <swift/AST/SourceFile.h>
-#include <swift/Basic/SourceManager.h>
-#include <llvm/Support/FileSystem.h>
+#include "swift/extractor/infra/SwiftTagTraits.h"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/SourceFile.h"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/Basic/SourceManager.h"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/llvm/Support/FileSystem.h"
 
 namespace codeql {
