+++ swift/extractor/infra/SwiftTagTraits.h
@@ -2,7 +2,7 @@
 
 // This file implements the mapping needed by the API defined in the TrapTagTraits.h, so that
 // TrapTagOf/TrapLabelOf provide the tags/labels for specific swift entity types.
-#include <swift/AST/ASTVisitor.h>
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/ASTVisitor.h"
 #include "swift/extractor/trap/TrapTagTraits.h"
 #include "swift/extractor/trap/generated/TrapTags.h"
 
@@ -39,32 +39,32 @@ MAP_TAG(StmtCondition);
 MAP_TAG(CaseLabelItem);
 #define ABSTRACT_STMT(CLASS, PARENT) MAP_SUBTAG(CLASS##Stmt, PARENT)
 #define STMT(CLASS, PARENT) ABSTRACT_STMT(CLASS, PARENT)
-#include "swift/AST/StmtNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/StmtNodes.def"
 
 MAP_TAG(Expr);
 #define ABSTRACT_EXPR(CLASS, PARENT) MAP_SUBTAG(CLASS##Expr, PARENT)
 #define EXPR(CLASS, PARENT) ABSTRACT_EXPR(CLASS, PARENT)
-#include "swift/AST/ExprNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/ExprNodes.def"
 
 MAP_TAG(Decl);
 #define ABSTRACT_DECL(CLASS, PARENT) MAP_SUBTAG(CLASS##Decl, PARENT)
 #define DECL(CLASS, PARENT) ABSTRACT_DECL(CLASS, PARENT)
-#include "swift/AST/DeclNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/DeclNodes.def"
 
 MAP_TAG(Pattern);
 #define ABSTRACT_PATTERN(CLASS, PARENT) MAP_SUBTAG(CLASS##Pattern, PARENT)
 #define PATTERN(CLASS, PARENT) ABSTRACT_PATTERN(CLASS, PARENT)
-#include "swift/AST/PatternNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/PatternNodes.def"
 
 MAP_TAG(TypeRepr);
 #define ABSTRACT_TYPEREPR(CLASS, PARENT) MAP_SUBTAG(CLASS##TypeRepr, PARENT)
 #define TYPEREPR(CLASS, PARENT) ABSTRACT_TYPEREPR(CLASS, PARENT)
-#include "swift/AST/TypeReprNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/TypeReprNodes.def"
 
 MAP_TYPE_TO_TAG(TypeBase, TypeTag);
 #define ABSTRACT_TYPE(CLASS, PARENT) MAP_SUBTAG(CLASS##Type, PARENT)
 #define TYPE(CLASS, PARENT) ABSTRACT_TYPE(CLASS, PARENT)
-#include "swift/AST/TypeNodes.def"
+#include "external/swift_prebuilt_linux/_virtual_includes/swift-llvm-support/swift/AST/TypeNodes.def"
 
 OVERRIDE_TAG(FuncDecl, ConcreteFuncDeclTag);
 OVERRIDE_TAG(VarDecl, ConcreteVarDeclTag);
