+++ swift/extractor/visitors/BUILD.bazel
@@ -1,0 +1,12 @@
+load("//swift:rules.bzl", "swift_cc_library")
+
+swift_cc_library(
+    name = "visitors",
+    srcs = glob(["*.cpp"]),
+    hdrs = glob(["*.h"]),
+    visibility = ["//swift:__subpackages__"],
+    deps = [
+        "//swift/extractor/infra",
+        "//swift/tools/prebuilt:swift-llvm-support",
+    ],
+)
