+++ swift/extractor/visitors/TypeReprVisitor.cpp
@@ -1,0 +1,3 @@
+#include "swift/extractor/visitors/TypeReprVisitor.h"
+
+namespace codeql {}  // namespace codeql
